Move platform specifics to their own file
